my_app/requirements.txt
Pillow
flask
numpy
torchvision